Remove unused middlware functions, add support for custom middleware
